rename macos example folder
