+++ MacOSExample/src/simple/draggable/index.tsx
