+++ example/App.tsx
@@ -75,12 +75,12 @@ import EmptyExample from './src/empty/EmptyExample';
 import RectButtonBorders from './src/release_tests/rectButton';
 import { ListWithHeader } from './src/ListWithHeader';
 
-import MacosDraggable from '../MacOSExample/src/basic/draggable';
+import MacosDraggable from '../MacOSExample/src/simple/draggable';
 import MacosPinchableBox from '../MacOSExample/src/recipes/scaleAndRotate';
-import Tap from '../MacOSExample/src/basic/tap';
-import LongPressExample from '../MacOSExample/src/basic/longPress';
-import ManualExample from '../MacOSExample/src/basic/manual';
-import HoverExample from '../MacOSExample/src/basic/hover';
+import Tap from '../MacOSExample/src/simple/tap';
+import LongPressExample from '../MacOSExample/src/simple/longPress';
+import ManualExample from '../MacOSExample/src/simple/manual';
+import HoverExample from '../MacOSExample/src/simple/hover';
 
 interface Example {
   name: string;
